move databaseClient out of own folder
