+++ server/modules/databaseClient.js
@@ -1,9 +1,9 @@
 // const storeTrade = require('./storeTrade');
 // const getUnsettledTrades = require('./getUnsettledTrades');
-const authedClient = require('../authedClient');
-const pool = require('../pool');
-const robot = require('../robot');
-const socketClient = require('../socketClient');
+const authedClient = require('./authedClient');
+const pool = require('./pool');
+const robot = require('./robot');
+const socketClient = require('./socketClient');
 
 
 const storeTrade = (newOrder, originalDetails) => {
